+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 * added factory function ``from_json`` to classes ``TimeProgPeriod``, ``TimeProgEntry`` and ``TimeProgram``
 * fixed issue with fault lists with larger number of entries (in ``HtHeatpump.get_fault_list()``);
   thanks to Alois for reporting
+* added new function ``HtParam.check_value_type`` to verify the correct type of a passed value
 
 1.2.0 (2019-06-10)
 ------------------